fix: Convert gallery page to static rendering
- Move gallery page to static rendering for better performance
- Separate artwork data into a dedicated file
- Remove client-side modal for simpler implementation
- Improve SEO and initial page load
